Cleaned up and sorted sample code.
