+++ lys_instr/tutorial/AdvEx2.py
@@ -1,46 +1,37 @@
-import pyqtgraph as pg
 from lys.widgets import LysSubWindow
 from lys.Qt import QtWidgets
-from lys_instr import DataStorage, gui, dummy
-
-
-class DetectorAdvEx2GUI(gui.MultiDetectorGUI):
-    def __init__(self, obj, wait=False, interval=1, iter=1):
-        super().__init__(obj, wait=wait, interval=interval, iter=iter)
-        if hasattr(self, "_expTime"):
-            self._expTime.setValue(0)
-            self._expTime.setEnabled(False)
-            
-    def _update(self):
-        if hasattr(self, "_data"):
-            self._data.axes = self._obj.axes[-2:]
-            self._mcut.cui.setRawWave(self._data)
-        if hasattr(self._obj, "_frame"):
-            self._framePlot.clear()
-            self._framePlot.plot(self._obj._frame[0], self._obj._frame[1])
+from lys_instr import DataStorage, PreCorrector, gui, dummy
+from lys_instr.gui.MultiScan import ScanWidget
 
 
 class AppWindow(LysSubWindow):
     def __init__(self, parent=None):
         super().__init__(parent)
-        self.setWindowTitle("Advanced Example #2")
+        # self.setWindowTitle("Advanced Example 2")
         self._storage = DataStorage()
-        self._detector = dummy.DetectorAdvEx2Dummy(indexShape=(36,), frameShape=(600,), exposure=0.1)
-        self._motor = dummy.MultiMotorDummy("x", "y")
+        self._detector = dummy.MultiDetectorDummy((5, 5), (256, 256), exposure=0.1)
+        self._motor = dummy.MultiMotorDummy("x", "y", "t")
+        self._pre = PreCorrector([self._motor])
+        self._switch = dummy.MultiSwitchDummy("A")
         self._storage.connect(self._detector)
+        # self.setSettingFile("AdvEx4.dic")
         self._initLayout()
-        self.setSettingFile("AdvEx2.dic")
+        
         self.adjustSize()
 
     def _initLayout(self):
         _storageGUI = gui.DataStorageGUI(self._storage)
-        _detectorGUI = DetectorAdvEx2GUI(self._detector)
-        _motorGUI = gui.MultiMotorGUI(self._motor)
-        _scanGUI = gui.MultiScan.ScanWidget(self._storage, [self._motor], {"DetectorAdvEx2Dummy": self._detector}, numScans=2)
+        _motorGUI = gui.MultiMotorGUI(self._motor, axisNamesSettable=("x", "y", "t"), axisNamesJoggable=("t"), axisNamesOffsettable=("x", "y", "t"))
+        _switchGUI = gui.MultiSwitchGUI(self._switch)
+        _detectorGUI = gui.MultiDetectorGUI(self._detector)
+        _correctorGUI = gui.PreCorrectorGUI(self._pre)
+        _scanGUI = ScanWidget(self._storage, [self._switch, self._motor], {"MultiDetectorDummy": self._detector}, numScans=2)
 
         self._tab = QtWidgets.QTabWidget()
         self._tab.addTab(_motorGUI, "Motor")
+        self._tab.addTab(_switchGUI, "Switch")
         self._tab.addTab(_scanGUI, "Scan")
+        self._tab.addTab(_correctorGUI, "PreCorr")
 
         VBox = QtWidgets.QVBoxLayout()
         VBox.addWidget(_storageGUI)
@@ -56,21 +47,18 @@ class AppWindow(LysSubWindow):
 
         # Set multicut display style
         mcut = _detectorGUI._mcut
-        graph1 = mcut.cui._children.addWave([1, 0])
-        mcut.display(graph1, type="grid", pos=(0, 0), wid=(3, 4))
-
-        # Add live frame display
-        self._framePlot = pg.PlotWidget()
-        mcut._grid.layout.addWidget(self._framePlot, 4, 0, 1, 4)
-        _detectorGUI._framePlot = self._framePlot
+        graph1 = mcut.cui._children.addWave([0, 1])
+        mcut.display(graph1, type="grid", pos=(0, 0), wid=(4, 2))
+        graph2 = mcut.cui._children.addWave([2, 3])
+        mcut.display(graph2, type="grid", pos=(0, 2), wid=(4, 2))
 
         # Set scan parameters
-        # Set scan axis 0 (first row) to 'y', linear mode, from 0, step 0.1, 9 steps
-        _scanGUI._scanRangeRows[0]._scanAxis.setCurrentText("y")
-        _scanGUI._scanRangeRows[0]._step.setValue(0.1)
-        _scanGUI._scanRangeRows[0]._numSteps.setValue(9)
-        # Set scan axis 1 (second row) to 'x', linear mode, from 0, step 0.1, 8 steps
-        _scanGUI._scanRangeRows[1]._scanAxis.setCurrentText("x")
+        _scanGUI._scanRangeRows[0]._scanAxis.setCurrentText("A")
+        _scanGUI._scanRangeRows[0]._scanMode.setCurrentText("Free")
+        _scanGUI._scanRangeRows[0]._freeExpr.setText(str([True,]))
+        _scanGUI._scanRangeRows[1]._scanAxis.setCurrentText("t")
         _scanGUI._scanRangeRows[1]._step.setValue(0.1)
-        _scanGUI._scanRangeRows[1]._numSteps.setValue(8)
-        _scanGUI._exposure.setValue(0.1)
+        _scanGUI._scanRangeRows[1]._numSteps.setValue(9)
+        _scanGUI._check_ref.setChecked(True)
+        _scanGUI._combo_ref.setCurrentText("A")      # Reference axis is "A"
+        _scanGUI._value_ref.setValue(False)          # Reference value is pump OFF
